+++ src/main/java/com/fairytaler/fairytalecat/community/command/application/dao/NoticeDAO.java
@@ -1,0 +1,2 @@
+package com.fairytaler.fairytalecat.community.command.application.dao;public class NoticeDAO {
+}
