+++ src/main/java/com/fairytaler/fairytalecat/community/command/application/service/NoticeService.java
@@ -1,0 +1,2 @@
+package com.fairytaler.fairytalecat.community.command.application.service;public class NoticeService {
+}
